+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 -   [Breaking] Remove `deleteAllCookies()`
 -   Add `setSession()` and `clearSession()` method inside `locals`
 -   Add `adapter.user` and `adapter.session` config
+-   Make `getSessionAndUserBySessionId()` adapter method optional
 -   [Fix] Sessions renewed inside hooks are properly being set
 
 ## v0.13.2